Moved the ScopeNameMember to from Domain area Controls area as it will be used by more then just the Domain.
